+++ Engine/src/Render/Vulkan/VulkanRenderer.h
@@ -56,7 +56,7 @@ public:
     void Update() override;
     void EndFrame() override;
     
-    void SendPushConstants(void* data, size_t size, Shader* shader, PushConstant pushConstant) const override;
+    void SendPushConstants(void* data, uint32_t size, Shader* shader, PushConstant pushConstant) const override;
     void BindVertexBuffers(RHIVertexBuffer* _vertexBuffer, RHIIndexBuffer* _indexBuffer) const override;
     void DrawVertex(RHIVertexBuffer* _vertexBuffer, RHIIndexBuffer* _indexBuffer) override;
     void DrawVertexSubMesh(RHIIndexBuffer* _indexBuffer, uint32_t startIndex, uint32_t indexCount) override;
